Add initial tests based on MSTest testing framework (#24)
Initial thought was that issue #24 was cause by using the MSTest testing framework. Although this was not the case, it does make sense to add tests based on MSTest testing framework.

Still needs cleanup (rename the existing XUnit test project), support for both 2.1 and 3.1 SDK and investigate whether it is possible to code share between MSTest and XUnit test project.